Renamed fitters submodules
